feat(codegen): reorg asm module as a folder
