--- codegen/src/asm.rs
@@ -1,264 +1,0 @@
-//! Low level assembler implementation for EVM.
-//!
-//! TODO: refactor this module with Result as outputs. (issue-21)
-
-use crate::{Buffer, Error, Result};
-use opcodes::{for_each_shanghai_operator, OpCode as _, ShangHai as OpCode};
-use tracing::trace;
-
-/// Low level assembler implementation for EVM.
-pub struct Assembler {
-    /// Buffer of the assembler.
-    buffer: Buffer,
-    /// Gas counter.
-    ///
-    /// This is used to calculate the gas cost of the generated code.
-    ///
-    /// TODO: use a more precise type, eq `u256`. (issue-20)
-    gas: u128,
-    /// Memory pointer, maximum 32, 64-bit words.
-    pub mp: u8,
-    /// Stack pointer, maximum 12 items.
-    sp: u8,
-}
-
-impl Assembler {
-    /// New assembler
-    pub fn new(sp: u8) -> Self {
-        Self {
-            buffer: Buffer::new(),
-            gas: 0,
-            mp: 0,
-            sp,
-        }
-    }
-
-    /// Buffer of the assembler.
-    pub fn buffer(&self) -> &[u8] {
-        &self.buffer
-    }
-
-    /// Mutable buffer of the assembler.
-    pub fn buffer_mut(&mut self) -> &mut Buffer {
-        &mut self.buffer
-    }
-
-    /// Increment the gas counter.
-    ///
-    /// TODO: use number bigger than `u256` for throwing proper errors. (#21)
-    pub fn increment_gas(&mut self, gas: u128) {
-        self.gas += gas;
-    }
-
-    /// Increment stack pointer
-    pub fn increment_sp(&mut self, items: u8) -> Result<()> {
-        self.sp += items;
-        if self.sp > 12 {
-            return Err(Error::StackOverflow(self.sp));
-        }
-
-        Ok(())
-    }
-
-    /// Decrement stack pointer
-    pub fn decrement_sp(&mut self, items: u8) -> Result<()> {
-        self.sp = self
-            .sp
-            .checked_sub(items)
-            .ok_or(Error::StackUnderflow(self.sp, items))?;
-        Ok(())
-    }
-
-    /// Increment memory pointer
-    pub fn increment_mp(&mut self, offset: u8) -> Result<()> {
-        self.mp += offset;
-        if self.mp > 32 {
-            return Err(Error::MemoryOutOfBounds);
-        }
-
-        Ok(())
-    }
-
-    /// Decrement memory pointer
-    pub fn decrement_mp(&mut self, offset: u8) -> Result<()> {
-        self.mp = self
-            .mp
-            .checked_sub(offset)
-            .ok_or(Error::MemoryOutOfBounds)?;
-        Ok(())
-    }
-
-    /// Emit a byte.
-    pub fn emit(&mut self, byte: u8) {
-        self.buffer.push(byte);
-    }
-
-    /// Emit n bytes.
-    pub fn emitn(&mut self, bytes: &[u8]) {
-        self.buffer.extend_from_slice(bytes);
-    }
-
-    /// Emit a single opcode.
-    ///
-    /// Mock the stack input and ouput for checking
-    /// the stack usages.
-    pub fn emit_op(&mut self, opcode: OpCode) -> Result<()> {
-        self.decrement_sp(opcode.stack_in() as u8)?;
-        self.emit(opcode.into());
-        self.increment_gas(opcode.gas().into());
-        self.increment_sp(opcode.stack_out() as u8)?;
-
-        Ok(())
-    }
-}
-
-macro_rules! impl_opcodes {
-    ($($name:ident => $opcode:ident),+) => {
-        $(
-            #[doc = concat!(" Emit", stringify!($opcode))]
-            pub fn $name(&mut self) -> Result<()> {
-                self.emit_op(OpCode::$opcode)?;
-                Ok(())
-            }
-        )*
-    };
-}
-
-/// Basic instruction implementations
-impl Assembler {
-    for_each_shanghai_operator!(impl_opcodes);
-
-    /// Load n bytes to extend self as another number type.
-    ///
-    /// Just for adapting the WASM instructions, this method makes
-    /// no sense for EVM since all of the numbers as U256.
-    pub(crate) fn _load(&mut self) -> Result<()> {
-        trace!("load");
-        Ok(())
-    }
-
-    /// Load 1 byte to extend self as another number type.
-    ///
-    /// Just for adapting the WASM instructions, this method makes
-    /// no sense for EVM since all of the numbers as U256.
-    pub(crate) fn _load8(&mut self) -> Result<()> {
-        trace!("load8");
-        Ok(())
-    }
-
-    /// Load 2 bytes to extend self as another number type.
-    ///
-    /// Just for adapting the WASM instructions, this method makes
-    /// no sense for EVM since all of the numbers as U256.
-    pub(crate) fn _load16(&mut self) -> Result<()> {
-        trace!("load16");
-        Ok(())
-    }
-
-    /// Load 4 bytes to extend self as another number type.
-    ///
-    /// Just for adapting the WASM instructions, this method makes
-    /// no sense for EVM since all of the numbers as U256.
-    pub(crate) fn _load32(&mut self) -> Result<()> {
-        trace!("load32");
-        Ok(())
-    }
-
-    /// Sign-agnostic compare unequal.
-    pub fn _ne(&mut self) -> Result<()> {
-        self.emit_op(OpCode::EQ)?;
-        self.emit_op(OpCode::ISZERO)?;
-        Ok(())
-    }
-
-    /// Simple not operator
-    pub fn _eqz(&mut self) -> Result<()> {
-        self.emit_op(OpCode::ISZERO)?;
-        Ok(())
-    }
-
-    /// Maximum of two values
-    pub fn _max(&mut self) -> Result<()> {
-        todo!()
-    }
-
-    /// Minium of two values
-    pub fn _min(&mut self) -> Result<()> {
-        todo!()
-    }
-
-    /// Ceiling operator
-    pub fn _ceil(&mut self) -> Result<()> {
-        todo!()
-    }
-
-    /// Floor operator
-    pub fn _floor(&mut self) -> Result<()> {
-        todo!()
-    }
-
-    /// Round to nearest integer, ties to even.
-    pub fn _nearest(&mut self) -> Result<()> {
-        todo!()
-    }
-
-    /// Square root
-    pub fn _sqrt(&mut self) -> Result<()> {
-        todo!()
-    }
-
-    /// Absolute value
-    pub fn _abs(&mut self) -> Result<()> {
-        todo!()
-    }
-
-    /// Negation
-    pub fn _neg(&mut self) -> Result<()> {
-        todo!()
-    }
-
-    /// If z1 and z2 have the same sign, return z1, otherwise
-    /// return z1 with negated sign.
-    pub fn _copysign(&mut self) -> Result<()> {
-        todo!()
-    }
-
-    // integer
-
-    /// sign-agnostic rotate left
-    ///
-    /// Return the result of rotating i1 left by k bits.
-    pub fn _rotl(&mut self) -> Result<()> {
-        todo!()
-    }
-
-    /// sign-agnostic rotate right
-    ///
-    /// Return the result of rotating i1 right by k bits.
-    pub fn _rotr(&mut self) -> Result<()> {
-        todo!()
-    }
-
-    /// sign-agnostic count leading zero bits
-    ///
-    /// Return the number of leading zero bits in i, all zero bits
-    /// are considered leading if the value is zero.
-    pub fn _clz(&mut self) -> Result<()> {
-        todo!()
-    }
-
-    /// sign-agnostic count leading zero bits
-    ///
-    /// Return the number of leading zero bits in i, all zero bits
-    /// are considered trailing if the value is zero.
-    pub fn _ctz(&mut self) -> Result<()> {
-        todo!()
-    }
-
-    /// sign-agnostic count number of one bits
-    ///
-    /// Return the count of no zero bits in i.
-    pub fn _popcnt(&mut self) -> Result<()> {
-        todo!()
-    }
-}
