+++ codegen/src/asm/float.rs
@@ -1,0 +1,51 @@
+//! Float Instructions
+
+use crate::{Assembler, Result};
+
+impl Assembler {
+    /// Maximum of two values
+    pub fn _max(&mut self) -> Result<()> {
+        todo!()
+    }
+
+    /// Minium of two values
+    pub fn _min(&mut self) -> Result<()> {
+        todo!()
+    }
+
+    /// Ceiling operator
+    pub fn _ceil(&mut self) -> Result<()> {
+        todo!()
+    }
+
+    /// Floor operator
+    pub fn _floor(&mut self) -> Result<()> {
+        todo!()
+    }
+
+    /// Round to nearest integer, ties to even.
+    pub fn _nearest(&mut self) -> Result<()> {
+        todo!()
+    }
+
+    /// Square root
+    pub fn _sqrt(&mut self) -> Result<()> {
+        todo!()
+    }
+
+    /// Absolute value
+    pub fn _abs(&mut self) -> Result<()> {
+        todo!()
+    }
+
+    /// Negation
+    pub fn _neg(&mut self) -> Result<()> {
+        todo!()
+    }
+
+    /// If z1 and z2 have the same sign, return z1, otherwise
+    /// return z1 with negated sign.
+    pub fn _copysign(&mut self) -> Result<()> {
+        todo!()
+    }
+}
